06_report_notification/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
db-dtypes==1.0.4
dominate==2.7.0
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
kaleido==0.2.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
plotly==5.11.0
proto-plus==1.22.1
protobuf==4.21.9
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.6
requests==2.28.1
rsa==4.9
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
tenacity==8.1.0
urllib3==1.26.12
